# backend/docs/swagger.yaml
basePath: /api/v1
consumes:
- application/json
definitions:
  github_com_FNTR3455234_FinTrack_backend_internal_modelos.AlertaPresupuesto:
    properties:
      categoria_id:
        type: string
      disponible:
        type: number
      estado:
        type: string
      gastado:
        type: number
      mensaje:
        type: string
      monto_limite:
        type: number
      nombre:
        type: string
      porcentaje_usado:
        type: number
    type: object
...
  github_com_FNTR3455234_FinTrack_backend_internal_modelos.PeticionTransaccion:
    properties:
      categoria_id:
        type: string
      cuenta_id:
        type: string
      descripcion:
        maxLength: 140
        minLength: 1
        type: string
      fecha:
        type: string
      monto:
        type: number
      notas:
        maxLength: 500
        type: string
      tipo:
        enum:
        - ingreso
        - gasto
        type: string
    required:
    - categoria_id
    - cuenta_id
    - descripcion
    - fecha
    - monto
    - tipo
    type: object
...
  github_com_FNTR3455234_FinTrack_backend_internal_modelos.Transaccion:
    properties:
      actualizado_en:
        type: string
      categoria_id:
        type: string
      creado_en:
        type: string
      cuenta_id:
        type: string
      descripcion:
        type: string
      fecha:
        type: string
      id:
        type: string
      monto:
        type: number
      notas:
        type: string
      tipo:
        type: string
    type: object
  github_com_FNTR3455234_FinTrack_backend_internal_modelos.TransaccionCreada:
    properties:
      actualizado_en:
        type: string
      alerta:
        $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_modelos.AlertaPresupuesto'
      categoria_id:
        type: string
      creado_en:
        type: string
      cuenta_id:
        type: string
      descripcion:
        type: string
      fecha:
        type: string
      id:
        type: string
      monto:
        type: number
      notas:
        type: string
      tipo:
        type: string
    type: object
...
  github_com_FNTR3455234_FinTrack_backend_internal_respuestas.Detalle:
    properties:
      codigo:
        type: string
      detalles:
        items:
          type: string
        type: array
      mensaje:
        type: string
    type: object
  github_com_FNTR3455234_FinTrack_backend_internal_respuestas.Meta:
    properties:
      limite:
        type: integer
      pagina:
        type: integer
      total:
        type: integer
      total_paginas:
        type: integer
    type: object
  github_com_FNTR3455234_FinTrack_backend_internal_respuestas.Sobre:
    properties:
      datos: {}
      meta:
        $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.Meta'
    type: object
  github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError:
    properties:
      error:
        $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.Detalle'
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: |-
    API REST de finanzas personales: cuentas, categorias, transacciones, presupuestos y reportes.
    Todos los datos pertenecen a un usuario. El identificador del dueño sale siempre del token, nunca del cuerpo de la peticion.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: FinTrack API
  version: 0.1.0
paths:
...
  /transacciones:
    get:
      description: Listado paginado con filtros. La meta trae pagina, limite, total
        y total_paginas.
      parameters:
      - description: Fecha inicial AAAA-MM-DD
        in: query
        name: desde
        type: string
      - description: Fecha final AAAA-MM-DD, incluida
        in: query
        name: hasta
        type: string
      - description: ingreso o gasto
        enum:
        - ingreso
        - gasto
        in: query
        name: tipo
        type: string
      - description: Filtra por categoria
        in: query
        name: categoria_id
        type: string
      - description: Filtra por cuenta
        in: query
        name: cuenta_id
        type: string
      - description: Texto en descripcion o notas
        in: query
        name: busqueda
        type: string
      - description: Orden del listado
        enum:
        - fecha_desc
        - fecha_asc
        - monto_desc
        - monto_asc
        in: query
        name: orden
        type: string
      - description: Pagina, desde 1
        in: query
        name: pagina
        type: integer
      - description: Resultados por pagina, de 1 a 100
        in: query
        name: limite
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.Sobre'
            - properties:
                datos:
                  items:
                    $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_modelos.Transaccion'
                  type: array
                meta:
                  $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.Meta'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
      security:
      - BearerAuth: []
      summary: Listar transacciones
      tags:
      - transacciones
    post:
      description: Si el gasto deja su categoria al 80% o mas de su presupuesto del
        mes, la respuesta trae ademas el campo alerta.
      parameters:
      - description: Datos del movimiento
        in: body
        name: cuerpo
        required: true
        schema:
          $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_modelos.PeticionTransaccion'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.Sobre'
            - properties:
                datos:
                  $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_modelos.TransaccionCreada'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
      security:
      - BearerAuth: []
      summary: Registrar un movimiento
      tags:
      - transacciones
  /transacciones/{id}:
    delete:
      description: 'Se borra de verdad: una transaccion no es referencia de nada mas.'
      parameters:
      - description: Identificador de 24 caracteres
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Sin contenido
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
      security:
      - BearerAuth: []
      summary: Borrar un movimiento
      tags:
      - transacciones
    get:
      parameters:
      - description: Identificador de 24 caracteres
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.Sobre'
            - properties:
                datos:
                  $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_modelos.Transaccion'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
      security:
      - BearerAuth: []
      summary: Obtener una transaccion
      tags:
      - transacciones
    put:
      description: El tipo tiene que seguir coincidiendo con el de su categoria.
      parameters:
      - description: Identificador de 24 caracteres
        in: path
        name: id
        required: true
        type: string
      - description: Datos del movimiento
        in: body
        name: cuerpo
        required: true
        schema:
          $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_modelos.PeticionTransaccion'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.Sobre'
            - properties:
                datos:
                  $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_modelos.Transaccion'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_FNTR3455234_FinTrack_backend_internal_respuestas.SobreError'
      security:
      - BearerAuth: []
      summary: Editar un movimiento
      tags:
      - transacciones
...
produces:
- application/json
securityDefinitions:
  BearerAuth:
    description: 'Token de acceso con el prefijo. Ejemplo: "Bearer eyJhbGciOi..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"